Remove build prefix files only on osx

+++ recipe/build.sh
@@ -18,10 +18,12 @@ if [[ "$cxx_compiler" == "gxx" ]]; then
   EXTRA_HOST_CLANG_FLAGS="-I$BUILD_PREFIX/$HOST/sysroot/usr/include"
 fi
 
-# avoid linking to libLLVM and libclang in build prefix. These are from the compiler package by anaconda
-rm -rf $BUILD_PREFIX/lib/libLLVM*.a $BUILD_PREFIX/lib/libclang*.a
-rm -rf $BUILD_PREFIX/include/llvm $BUILD_PREFIX/include/llvm-c
-rm -rf $BUILD_PREFIX/include/clang $BUILD_PREFIX/include/clang-c
+if [[ "$(uname)" == "Darwin" ]]; then
+    # avoid linking to libLLVM and libclang in build prefix. These are from the compiler package by anaconda
+    rm -rf $BUILD_PREFIX/lib/libLLVM*.a $BUILD_PREFIX/lib/libclang*.a
+    rm -rf $BUILD_PREFIX/include/llvm $BUILD_PREFIX/include/llvm-c
+    rm -rf $BUILD_PREFIX/include/clang $BUILD_PREFIX/include/clang-c
+fi
 
 if [[ "$(uname)" == "Darwin" || "$c_compiler" == "toolchain_c" ]]; then
   export CC=$PREFIX/bin/clang
